cleanup
git-svn-id: http://svn.code.sf.net/p/netpbm/code/trunk@3235 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ editor/pamflip/pamflip.c
@@ -222,8 +222,8 @@ xformDimensions(struct XformCore const xform,
    E.g. if it's a 90 degree rotation of a 10 x 20 image, the output is
    a 20 x 10 image.
 -----------------------------------------------------------------------------*/
-    *outColsP = abs(xform.a * inCols + xform.c * inRows);
-    *outRowsP = abs(xform.b * inCols + xform.d * inRows);
+    *outColsP = abs(xform.a * (int)inCols + xform.c * (int)inRows);
+    *outRowsP = abs(xform.b * (int)inCols + xform.d * (int)inRows);
 }
